refactor(server): split checkout and order-report routes into dedicated modules
