+++ server/app/src/main/kotlin/pos/ambrosia/api/Orders.kt
@@ -15,91 +15,25 @@ import pos.ambrosia.db.DatabaseConnection
 import pos.ambrosia.logger
 import pos.ambrosia.models.AddOrderDishRequest
 import pos.ambrosia.models.CompleteOrder
-import pos.ambrosia.models.Message
 import pos.ambrosia.models.Order
 import pos.ambrosia.models.OrderDish
 import pos.ambrosia.models.OrderWithDishesRequest
-import pos.ambrosia.models.OrderWithPaymentFilters
 import pos.ambrosia.services.OrderService
-import pos.ambrosia.services.ReportService
 import pos.ambrosia.utils.DatabaseException
 import pos.ambrosia.utils.ResourceNotFoundException
 import pos.ambrosia.utils.authorizePermission
 import java.sql.Connection
-import java.time.LocalDate
-
-private fun parseDateQueryParam(
-    value: String?,
-    name: String,
-): String? {
-    if (value.isNullOrBlank()) return null
-    return try {
-        LocalDate.parse(value).toString()
-    } catch (_: Exception) {
-        throw IllegalArgumentException("Invalid $name: $value. Expected format YYYY-MM-DD")
-    }
-}
-
-private fun parseDoubleQueryParam(
-    value: String?,
-    name: String,
-): Double? {
-    if (value.isNullOrBlank()) return null
-    return value.toDoubleOrNull() ?: throw IllegalArgumentException("Invalid $name: $value")
-}
 
 fun Application.configureOrders() {
     val connection: Connection = DatabaseConnection.getConnection()
     val orderService = OrderService(connection)
-    val reportService = ReportService(connection)
-    routing { route("/orders") { orders(orderService, reportService) } }
+    routing { route("/orders") { orders(orderService) } }
 }
 
-fun Route.orders(
-    orderService: OrderService,
-    reportService: ReportService,
-) {
+fun Route.orders(orderService: OrderService) {
     authorizePermission("orders_read") {
         get("") {
             val orders = orderService.getOrders()
-            if (orders.isEmpty()) {
-                call.respond(HttpStatusCode.OK, "No orders found")
-                return@get
-            }
-            call.respond(HttpStatusCode.OK, orders)
-        }
-
-        get("/with-payments") {
-            val filters =
-                try {
-                    OrderWithPaymentFilters(
-                        startDate = parseDateQueryParam(call.request.queryParameters["startDate"], "startDate"),
-                        endDate = parseDateQueryParam(call.request.queryParameters["endDate"], "endDate"),
-                        status = call.request.queryParameters["status"]?.takeIf { it.isNotBlank() },
-                        userId = call.request.queryParameters["userId"]?.takeIf { it.isNotBlank() },
-                        paymentMethod = call.request.queryParameters["paymentMethod"]?.takeIf { it.isNotBlank() },
-                        minTotal = parseDoubleQueryParam(call.request.queryParameters["minTotal"], "minTotal"),
-                        maxTotal = parseDoubleQueryParam(call.request.queryParameters["maxTotal"], "maxTotal"),
-                        sortBy = call.request.queryParameters["sortBy"]?.takeIf { it.isNotBlank() },
-                        sortOrder = call.request.queryParameters["sortOrder"]?.takeIf { it.isNotBlank() },
-                    ).also {
-                        if (it.startDate != null && it.endDate != null && it.startDate > it.endDate) {
-                            throw IllegalArgumentException("startDate cannot be greater than endDate")
-                        }
-                    }
-                } catch (error: IllegalArgumentException) {
-                    call.respond(HttpStatusCode.BadRequest, Message(error.message ?: "Invalid query parameters"))
-                    return@get
-                }
-
-            val orders =
-                try {
-                    reportService.getOrdersWithPaymentsFiltered(filters)
-                } catch (error: IllegalArgumentException) {
-                    call.respond(HttpStatusCode.BadRequest, Message(error.message ?: "Invalid query parameters"))
-                    return@get
-                }
-
             if (orders.isEmpty()) {
                 call.respond(HttpStatusCode.OK, "No orders found")
                 return@get
@@ -226,17 +160,6 @@ fun Route.orders(
                 return@get
             }
             call.respond(HttpStatusCode.OK, orders)
-        }
-
-        get("/total-sales/{date}") {
-            val date = call.parameters["date"]
-            if (date.isNullOrEmpty()) {
-                call.respond(HttpStatusCode.BadRequest, "Missing or malformed date")
-                return@get
-            }
-
-            val totalSales = reportService.getTotalSalesByDate(date)
-            call.respond(HttpStatusCode.OK, mapOf("date" to date, "total_sales" to totalSales.toString()))
         }
     }
